py/cli/requirements.txt
eywa-client==0.2.1
nanoid==2.0.0
numpy==2.2.0
